refactor: move and rename files for clarity
- `src/` now only contains the most relevant parts of Probot. `manfest-creation.ts` will most likely go into a separate packge for v11
- `src/octokit` has everything Octokit-related
- `src/server` has everything express related
- `src/helpers` everything else :)
