--- src/github/version.ts
@@ -1,1 +1,0 @@
-export const VERSION = "0.0.0-development";
