+++ src/helpers/get-private-key.ts
