+++ src/manifest-creation.ts
@@ -2,7 +2,7 @@ import fs from "fs";
 import yaml from "js-yaml";
 import path from "path";
 import updateDotenv from "update-dotenv";
-import { ProbotOctokit } from "./github/octokit";
+import { ProbotOctokit } from "./octokit/probot-octokit";
 
 export class ManifestCreation {
   get pkg() {
