+++ src/server/create-server.ts
@@ -2,7 +2,7 @@ import path from "path";
 
 import express from "express";
 
-import { getLoggingMiddleware } from "./middleware/logging";
+import { getLoggingMiddleware } from "./logging-middleware";
 
 import type { Logger } from "pino";
 
@@ -16,11 +16,11 @@ export const createServer = (options: ServerOptions) => {
   app.use(getLoggingMiddleware(options.logger));
   app.use(
     "/probot/static/",
-    express.static(path.join(__dirname, "..", "static"))
+    express.static(path.join(__dirname, "..", "..", "static"))
   );
   app.use(options.webhook);
   app.set("view engine", "hbs");
-  app.set("views", path.join(__dirname, "..", "views"));
+  app.set("views", path.join(__dirname, "..", "..", "views"));
   app.get("/ping", (req, res) => res.end("PONG"));
 
   return app;
