+++ src/version.ts
@@ -1,0 +1,2 @@
+// The version is set automatically before publish to npm
+export const VERSION = "0.0.0-development";
