+++ test/resolver.test.ts
@@ -1,4 +1,7 @@
-import { resolve, Resolver } from "../src/resolver";
+import {
+  resolveAppFunction,
+  Resolver,
+} from "../src/helpers/resolve-app-function";
 
 const stubAppFnPath = require.resolve("./fixtures/plugin/stub-plugin");
 const basedir = process.cwd();
@@ -11,7 +14,7 @@ describe("resolver", () => {
   });
 
   it("loads the module at the resolved path", () => {
-    const module = resolve("foo", { resolver: stubResolver });
+    const module = resolveAppFunction("foo", { resolver: stubResolver });
     expect(module).toBe(require(stubAppFnPath));
     expect(stubResolver).toHaveBeenCalledWith("foo", { basedir });
   });
